backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
graphiti-core
anthropic>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
neo4j>=5.0.0
